refactor: united ImportShopProductsFileStack
